requirements_clean.txt.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
numpy>=1.24.0
scipy==1.15.3
scikit-learn==1.7.2
torch==2.5.1
transformers==4.57.3
sentence-transformers==5.2.0
bertopic==0.15.0
umap-learn==0.5.9.post2
hdbscan==0.8.41
faiss-cpu==1.7.4
google-generativeai==0.3.0
nltk==3.8.1
plotly==6.5.0
